requirements.txt
fastapi
uvicorn
python-multipart
pefile
yara-python
colorama
